Abandoning certain functions and fixes for GH runner:
Abandoning the dir_copy_if_exists function because we are trying to copy <folder>/*, where
bash resolves the * to all the files in the folder and the function ends up with:
  dir_copy_if_exists file1 file2 .... <relevant arguments>

Therefore, only using 'cp -r' and '*'.

Setting all the user/group permissions at the end, once for all the folder structure.

Temp. commenting the /dev/shm/performanceanalyzer, as it is failing to install from local
snap file.