[DPE-6851] Repurpose bin-wrapper.sh to call opensearch class
There is a need to now and then call opensearch without starting the service.

However, we may have set OpenSearch to consume most of the node's RAM, and if we naively
use the same jvm.options, then we will end up with a OOM as both java processes will
compete for memory.

This PR repurposes the bin-wrapper.sh: this script manually calls the opensearch class
and enforces a reasonable RAM consumption. It should be used for smaller checks such as
asking for OpenSearch version. There is a new script that replaces original bin-wrapper.sh:
cmd-wrapper.sh.